+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+* fix urls ending with slash [#365](https://github.com/jcubic/jquery.terminal/issues/365)
+
 ## 1.11.4
 * handle non string and functions in error the same as in echo
 * fix selection for raw output (reported by @ovk)