+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.12.0
 
 ### Features
 * default options for cmd plugin
@@ -7,7 +7,8 @@
 ### Bugfixes
 * fix urls ending with slash [#365](https://github.com/jcubic/jquery.terminal/issues/365)
 * stringify non string commands in set_command
-* fix scrolling of the page after you click on the link
+* fix scrolling of the page, when press space, after you click on the link
+* fix scrolling flicker when terminal added to body
 
 ## 1.11.4
 * handle non string and functions in error the same as in echo